# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/push-token": {
      "post": {
        "operationId": "AuthController_savePushToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTokenResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_removePushToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemovePushTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/appointments": {
      "get": {
        "operationId": "AppointmentsController_findAll",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "operationId": "AppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/me": {
      "get": {
        "operationId": "AppointmentsController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyAppointmentsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/availability": {
      "get": {
        "operationId": "AppointmentsController_getAvailability",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "09:00"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{id}/reschedule": {
      "patch": {
        "operationId": "AppointmentsController_reschedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{id}": {
      "delete": {
        "operationId": "AppointmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{id}/status": {
      "patch": {
        "operationId": "AppointmentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/recurring": {
      "get": {
        "operationId": "AppointmentsController_findRecurring",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecurringBookingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "operationId": "AppointmentsController_createRecurring",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringBookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/recurring/{id}": {
      "patch": {
        "operationId": "AppointmentsController_updateRecurring",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringBookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      },
      "delete": {
        "operationId": "AppointmentsController_removeRecurring",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "appointments"
        ]
      }
    },
...
  },
  "info": {
    "title": "BarberTchê API",
    "description": "Contrato HTTP do backend BarberTchê.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "passwordConfirmation": {
            "type": "string",
            "minLength": 8
          },
          "termsAccepted": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "phone",
          "email",
          "password",
          "passwordConfirmation",
          "termsAccepted"
        ]
      },
      "ClientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone"
        ]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CLIENT"
            ]
          },
          "client": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ClientResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "role",
          "client"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AuthMeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "PushTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          },
          "deviceId": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
      "PushTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          },
          "deviceId": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "platform",
          "updatedAt"
        ]
      },
      "RemovePushTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512
          }
        },
        "required": [
          "token"
        ]
      },
...
      "DeletedResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "AppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "start": {
            "type": "string",
            "example": "09:00"
          },
          "end": {
            "type": "string",
            "example": "09:30"
          },
          "clientId": {
            "type": "string",
            "nullable": true
          },
          "clientName": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "app",
              "manual",
              "recurring"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELED",
              "NO_SHOW"
            ]
          },
          "canceledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "date",
          "start",
          "end",
          "clientName",
          "serviceId",
          "source",
          "status"
        ]
      },
      "RecurringBookingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "weekday": {
            "type": "number",
            "minimum": 0,
            "maximum": 6
          },
          "start": {
            "type": "string",
            "example": "09:00"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "clientId",
          "serviceId",
          "weekday",
          "start",
          "active"
        ]
      },
      "MyAppointmentsResponseDto": {
        "type": "object",
        "properties": {
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentResponseDto"
            }
          },
          "recurring": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecurringBookingResponseDto"
            }
          }
        },
        "required": [
          "appointments",
          "recurring"
        ]
      },
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "app",
              "manual"
            ]
          }
        },
        "required": [
          "date",
          "start",
          "serviceId"
        ]
      },
      "RescheduleAppointmentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "start": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "start"
        ]
      },
      "AppointmentStatus": {
        "type": "string",
        "enum": [
          "SCHEDULED",
          "COMPLETED",
          "CANCELED",
          "NO_SHOW"
        ]
      },
      "UpdateAppointmentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AppointmentStatus"
              }
            ]
          },
          "cancellationReason": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateRecurringBookingDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "weekday": {
            "type": "number",
            "minimum": 0,
            "maximum": 6
          },
          "start": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "clientId",
          "serviceId",
          "weekday",
          "start"
        ]
      },
...
    }
  }
}